Move HTTP logic to separate folder
