+++ source/Databricks/source/SqlStatementExecution.UnitTests/Extensions/DependencyInjection/DatabricksSqlStatementExecutionExtensionsTests.cs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Http;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
