+++ source/Databricks/source/SqlStatementExecution/Http/HttpClientNameConstants.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution;
+namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.Http;
 
 /// <summary>
 /// Constants used for naming <see cref="HttpClient"/> instances.
